+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/repository/EtapaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.utfpr.pb.projetotcc.model.Etapa;
 
 public interface EtapaRepository
-			extends JpaRepository<Etapa, Integer>{
+			extends JpaRepository<Etapa, Long>{
 
 }
