+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/service/EtapaService.java
@@ -3,6 +3,6 @@ package br.edu.utfpr.pb.projetotcc.service;
 import br.edu.utfpr.pb.projetotcc.model.Etapa;
 
 public interface EtapaService
-			extends CrudService<Etapa, Integer>{
+			extends CrudService<Etapa, Long>{
 
 }
